[SERVER-MODULE] - Evoluindo segurança.

+++ server-libs/server-commons/src/main/java/br/net/mirante/singular/server/commons/flow/rest/DefaultServerREST.java
@@ -5,7 +5,10 @@
 
 package br.net.mirante.singular.server.commons.flow.rest;
 
+import static br.net.mirante.singular.server.commons.util.ServerActionConstants.ACTION_DELETE;
+
 import javax.inject.Inject;
+import javax.inject.Named;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -16,13 +19,18 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import br.net.mirante.singular.flow.core.ProcessDefinition;
+import br.net.mirante.singular.form.context.SFormConfig;
 import br.net.mirante.singular.form.spring.SpringServiceRegistry;
-import br.net.mirante.singular.server.commons.config.SingularServerConfiguration;
 import br.net.mirante.singular.server.commons.persistence.entity.form.PetitionEntity;
 import br.net.mirante.singular.server.commons.service.PetitionService;
+import br.net.mirante.singular.server.commons.spring.security.PermissionResolverService;
 import br.net.mirante.singular.server.commons.util.PetitionUtil;
 import br.net.mirante.singular.support.spring.util.AutoScanDisabled;
 
+/**
+ * Essa interface deve ser protegida de forma que apenas o próprio servidor possa
+ * acessar seus métodos.
+ */
 @AutoScanDisabled
 @RequestMapping("/rest/flow")
 @RestController
@@ -38,22 +46,31 @@ public class DefaultServerREST {
     protected PetitionService<PetitionEntity> petitionService;
 
     @Inject
-    private SingularServerConfiguration singularServerConfiguration;
+    private PermissionResolverService permissionResolverService;
 
     @Inject
     private SpringServiceRegistry springServiceRegistry;
 
+    @Inject
+    @Named("formConfigWithDatabase")
+    protected SFormConfig<String> singularFormConfig;
+
     @RequestMapping(value = PATH_BOX_ACTION + DELETE, method = RequestMethod.POST)
-    public ActionResponse excluir(@RequestBody Long id) {
+    public ActionResponse excluir(@RequestParam Long id, @RequestBody Action action) {
         try {
-            petitionService.delete(id);
+            boolean hasPermission = permissionResolverService.hasPermission(id, action.getIdUsuario(), ACTION_DELETE);
+            if (hasPermission) {
+                petitionService.delete(id);
+                return new ActionResponse("Registro excluído com sucesso", true);
+            } else {
+                return new ActionResponse("Você não tem permissão para executar esta ação.", false);
+            }
         } catch (Exception e) {
             final String msg = "Erro ao excluir o item.";
             LOGGER.error(msg, e);
             return new ActionResponse(msg, false);
         }
 
-        return new ActionResponse("Registro excluído com sucesso", true);
     }
 
     @RequestMapping(value = PATH_BOX_ACTION + EXECUTE, method = RequestMethod.POST)
@@ -63,7 +80,7 @@ public class DefaultServerREST {
             final ProcessDefinition<?> processDefinition = PetitionUtil.getProcessDefinition(petition);
 
             IController controller = getActionController(processDefinition, action);
-            return controller.execute(petition, action);
+            return controller.run(petition, action);
         } catch (Exception e) {
             final String msg = String.format("Erro ao executar a ação %s para o id %d.", action.getName(), id);
             LOGGER.error(msg, e);
